`atreus`: restore intended matrix implementations (#24082)
* `atreus`: restore intended matrix implementations

PR 24010 inadvertently changed the switch matrix for the
`atreus/astar_mirrored` keyboard in removing the `PCBDOWN` configuration
option, and also made PCB up and down implementations available for all
Atreus variants, which is not intended usage.

This commit restores the intended behaviour, and makes all the keymaps
which are currently present agnostic to hardware variant.